+++ persistent-mongoDB/test/EmbedTestMongo.hs
@@ -20,7 +20,6 @@ import Data.List.NonEmpty hiding (insert, length)
 import qualified Data.Map as M
 import qualified Data.Set as S
 import qualified Data.Text as T
-import Data.Typeable (Typeable)
 import Database.MongoDB (genObjectId)
 import Database.MongoDB (Value(String))
 import System.Process (readProcess)
@@ -29,7 +28,7 @@ import EntityEmbedTestMongo
 import Database.Persist.MongoDB
 
 data TestException = TestException
-    deriving (Show, Typeable, Eq)
+    deriving (Show, Eq)
 instance Exception TestException
 
 instance PersistFieldSql a => PersistFieldSql (NonEmpty a) where